# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RentBox Energy API",
    "description": "API дашборда ВЭС: исторические прогнозы, состояние агента и CSV. Расчёты доступны после подключения агента и подтверждения настроек времени.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/agent/runs": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_agent_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastRunList"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_agent_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/agent/runs/{run_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_agent_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastRunRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/agent/runs/{run_id}/forecast": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Forecast",
        "operationId": "get_forecast_api_agent_runs__run_id__forecast_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/agent/runs/{run_id}/forecast.csv": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Csv",
        "operationId": "get_csv_api_agent_runs__run_id__forecast_csv_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/agent/replays": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Create Replay",
        "operationId": "create_replay_api_agent_replays_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayAccepted"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/replays/{replay_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Replay",
        "operationId": "get_replay_api_agent_replays__replay_id__get",
        "parameters": [
          {
            "name": "replay_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Replay Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "stage": {
            "type": "string",
            "enum": [
              "validate",
              "weather",
              "prepare",
              "model",
              "predict",
              "review",
              "save"
            ],
            "title": "Stage"
          },
          "level": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "title": "Level"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Attempt",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "stage",
          "level",
          "message"
        ],
        "title": "AgentEvent"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "ForecastAnalysis": {
        "properties": {
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "ForecastAnalysis"
      },
      "ForecastPoint": {
        "properties": {
          "valid_time": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Time"
          },
          "lead_hour": {
            "type": "integer",
            "maximum": 48.0,
            "minimum": 1.0,
            "title": "Lead Hour"
          },
          "predicted_power": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Predicted Power"
          },
          "weather_inputs": {
            "items": {
              "$ref": "#/components/schemas/WeatherInput"
            },
            "type": "array",
            "maxItems": 16,
            "minItems": 1,
            "title": "Weather Inputs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valid_time",
          "lead_hour",
          "predicted_power",
          "weather_inputs"
        ],
        "title": "ForecastPoint"
      },
      "ForecastRead": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "horizon_hours": {
            "type": "integer",
            "enum": [
              24,
              48
            ],
            "title": "Horizon Hours"
          },
          "unit": {
            "type": "string",
            "const": "normalized_power",
            "title": "Unit",
            "default": "normalized_power"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "training_data_available_until": {
            "type": "string",
            "format": "date-time",
            "title": "Training Data Available Until"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ForecastSeries"
            },
            "type": "array",
            "title": "Series"
          },
          "analysis": {
            "$ref": "#/components/schemas/ForecastAnalysis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "as_of",
          "horizon_hours",
          "model_version",
          "training_data_available_until",
          "series",
          "analysis"
        ],
        "title": "ForecastRead"
      },
      "ForecastRunCreate": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "horizon_hours": {
            "type": "integer",
            "enum": [
              24,
              48
            ],
            "title": "Horizon Hours",
            "default": 48
          },
          "turbine_ids": {
            "items": {
              "type": "integer",
              "maximum": 2.0,
              "minimum": 1.0
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 1,
            "title": "Turbine Ids"
          },
          "refresh_weather": {
            "type": "boolean",
            "title": "Refresh Weather",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of"
        ],
        "title": "ForecastRunCreate"
      },
      "ForecastRunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastRunRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ForecastRunList"
      },
      "ForecastRunRead": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "default": "queued"
          },
          "stage": {
            "type": "string",
            "enum": [
              "validate",
              "weather",
              "prepare",
              "model",
              "predict",
              "review",
              "save"
            ],
            "title": "Stage",
            "default": "validate"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          },
          "agent_mode": {
            "type": "string",
            "enum": [
              "llm",
              "policy"
            ],
            "title": "Agent Mode"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "horizon_hours": {
            "type": "integer",
            "enum": [
              24,
              48
            ],
            "title": "Horizon Hours"
          },
          "turbine_ids": {
            "items": {
              "type": "integer",
              "maximum": 2.0,
              "minimum": 1.0
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 1,
            "title": "Turbine Ids"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision",
            "default": 1
          },
          "supersedes_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Run Id"
          },
          "reused_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reused Run Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/AgentEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "agent_mode",
          "as_of",
          "horizon_hours",
          "turbine_ids"
        ],
        "title": "ForecastRunRead"
      },
      "ForecastSeries": {
        "properties": {
          "turbine_id": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Turbine Id"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            },
            "type": "array",
            "maxItems": 48,
            "minItems": 24,
            "title": "Points"
          },
          "weather": {
            "$ref": "#/components/schemas/WeatherProvenance"
          },
          "storage_run_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Storage Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turbine_id",
          "points",
          "weather",
          "storage_run_id"
        ],
        "title": "ForecastSeries"
      },
...
      "ReplayAccepted": {
        "properties": {
          "replay_id": {
            "type": "string",
            "title": "Replay Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "replay_id"
        ],
        "title": "ReplayAccepted"
      },
      "ReplayCreate": {
        "properties": {
          "first_as_of": {
            "type": "string",
            "format": "date-time",
            "title": "First As Of"
          },
          "last_as_of": {
            "type": "string",
            "format": "date-time",
            "title": "Last As Of"
          },
          "step_hours": {
            "type": "integer",
            "const": 24,
            "title": "Step Hours",
            "default": 24
          },
          "horizon_hours": {
            "type": "integer",
            "enum": [
              24,
              48
            ],
            "title": "Horizon Hours",
            "default": 48
          },
          "turbine_ids": {
            "items": {
              "type": "integer",
              "maximum": 2.0,
              "minimum": 1.0
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 1,
            "title": "Turbine Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "first_as_of",
          "last_as_of"
        ],
        "title": "ReplayCreate"
      },
      "ReplayRead": {
        "properties": {
          "replay_id": {
            "type": "string",
            "title": "Replay Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "default": "queued"
          },
          "total_runs": {
            "type": "integer",
            "title": "Total Runs"
          },
          "completed_runs": {
            "type": "integer",
            "title": "Completed Runs",
            "default": 0
          },
          "failed_runs": {
            "type": "integer",
            "title": "Failed Runs",
            "default": 0
          },
          "run_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Run Ids"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "replay_id",
          "total_runs",
          "run_ids"
        ],
        "title": "ReplayRead"
      },
      "RunAccepted": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunAccepted"
      },
...
      "WeatherProvenance": {
        "properties": {
          "sources": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SingleRunWeatherSource"
                },
                {
                  "$ref": "#/components/schemas/PreviousRunsWeatherSource"
                }
              ],
              "discriminator": {
                "propertyName": "product",
                "mapping": {
                  "previous_runs": "#/components/schemas/PreviousRunsWeatherSource",
                  "single_run": "#/components/schemas/SingleRunWeatherSource"
                }
              }
            },
            "type": "array",
            "maxItems": 128,
            "minItems": 1,
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sources"
        ],
        "title": "WeatherProvenance"
      }
    }
  }
}